# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/coupons": {
            "get": {
                "description": "ListController for Coupons",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "ListController",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginatedResponse-coupons_Coupon"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateController for Coupons",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coupons.CreateCouponRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-coupons_Coupon"
                        }
                    }
                }
            }
        },
...
        "/api/v1/coupons/{id}": {
            "delete": {
                "description": "DeleteController for Coupons",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "DeleteController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateController for Coupons",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Coupons"
                ],
                "summary": "UpdateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coupons.UpdateCouponRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-coupons_Coupon"
                        }
                    }
                }
            }
        },
        "/api/v1/courses": {
            "get": {
                "description": "ListController for Courses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "ListController",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.PaginatedResponse-courses_CourseCardResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateController for Courses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "CreateController",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.CreateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-courses_CourseCreatedResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/courses/{id}": {
            "delete": {
                "description": "DeleteController for Courses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "DeleteController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-utils_DeleteResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "UpdateController for Courses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "UpdateController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.UpdateCourseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-courses_Course"
                        }
                    }
                }
            }
        },
        "/api/v1/courses/{id}/study": {
            "get": {
                "description": "ReadStudyController for Courses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "ReadStudyController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-enrollments_CourseStudyResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/courses/{slug}": {
            "get": {
                "description": "ReadLandingController for Courses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "ReadLandingController",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse-courses_CourseLandingResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "coupons.Coupon": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "course": {
                    "$ref": "#/definitions/models.CourseInfo"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "discount_percent": {
                    "type": "number"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_usage": {
                    "type": "integer"
                },
                "usage_count": {
                    "type": "integer"
                }
            }
        },
...
        "coupons.CreateCouponRequest": {
            "type": "object",
            "required": [
                "code",
                "discount_percent",
                "expires_at",
                "max_usage"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "course_id": {
                    "type": "string"
                },
                "discount_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 1
                },
                "expires_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_usage": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "coupons.UpdateCouponRequest": {
            "type": "object",
            "properties": {
                "discount_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 1
                },
                "expires_at": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_usage": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "courses.Course": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number"
                },
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coupon_allowed": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "feedback_count": {
                    "type": "integer"
                },
                "final_price": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "rating_avg": {
                    "type": "number"
                },
                "requirements": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_duration_seconds": {
                    "type": "integer"
                },
                "total_lectures": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "courses.CourseCreatedResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "courses.CourseLandingResponse": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number"
                },
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category": {
                    "$ref": "#/definitions/models.CategoryInfo"
                },
                "chapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/courses.ChapterCardResponse"
                    }
                },
                "feedback_count": {
                    "type": "integer"
                },
                "final_price": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "instructor": {
                    "$ref": "#/definitions/models.InstructorInfo"
                },
                "is_enrolled": {
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "level": {
                    "type": "string"
                },
                "long_description": {
                    "type": "string"
                },
                "preview_video_url": {
                    "type": "string"
                },
                "rating_avg": {
                    "type": "number"
                },
                "requirements": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "short_description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "subcategory": {
                    "$ref": "#/definitions/models.CategoryInfo"
                },
                "title": {
                    "type": "string"
                },
                "total_duration_seconds": {
                    "type": "integer"
                },
                "total_lectures": {
                    "type": "integer"
                }
            }
        },
        "courses.CreateCourseRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "level": {
                    "type": "string",
                    "enum": [
                        "beginner",
                        "intermediate",
                        "advanced",
                        "all"
                    ]
                },
                "short_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ]
                },
                "subcategory_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 3
                }
            }
        },
...
        "courses.UpdateCourseRequest": {
            "type": "object",
            "properties": {
                "actual_price": {
                    "type": "number",
                    "minimum": 0
                },
                "benefits": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "category_id": {
                    "type": "string"
                },
                "coupon_allowed": {
                    "type": "boolean"
                },
                "final_price": {
                    "type": "number",
                    "minimum": 0
                },
                "image_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "level": {
                    "type": "string",
                    "enum": [
                        "beginner",
                        "intermediate",
                        "advanced",
                        "all"
                    ]
                },
                "long_description": {
                    "type": "string"
                },
                "preview_video_url": {
                    "type": "string"
                },
                "requirements": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "short_description": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ]
                },
                "subcategory_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 3
                }
            }
        },
...
        "enrollments.CourseStudyResponse": {
            "type": "object",
            "properties": {
                "chapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enrollments.StudyChapterItem"
                    }
                },
                "course": {
                    "$ref": "#/definitions/models.CourseInfo"
                },
                "enrollment": {
                    "$ref": "#/definitions/enrollments.EnrollmentStudyInfo"
                }
            }
        },
...
        "utils.DeleteResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "utils.PaginatedData-coupons_Coupon": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/coupons.Coupon"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "utils.PaginatedData-courses_CourseCardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/courses.CourseCardResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "utils.PaginatedResponse-coupons_Coupon": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.PaginatedData-coupons_Coupon"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.PaginatedResponse-courses_CourseCardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.PaginatedData-courses_CourseCardResponse"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-coupons_Coupon": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/coupons.Coupon"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-courses_Course": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/courses.Course"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SwaggerResponse-courses_CourseCreatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/courses.CourseCreatedResponse"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "utils.SwaggerResponse-courses_CourseLandingResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/courses.CourseLandingResponse"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-enrollments_CourseStudyResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/enrollments.CourseStudyResponse"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.SwaggerResponse-utils_DeleteResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.DeleteResponse"
                },
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}